+++ app/(main)/deal/components/DealNavbar.tsx
