+++ app/(main)/deal/hooks/useDealTimer.tsx
