+++ app/product/components/AddToCartForm.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import useAddToCartForm from "@/app/product/[id]/hooks/forms/useAddToCartForm";
+import useAddToCartForm from "@/app/product/hooks/forms/useAddToCartForm";
 
-import ProductOptions from "@/app/product/[id]/components/ProductOptions";
-import ProductQuantity from "@/app/product/[id]/components/ProductQuantity";
+import ProductOptions from "@/app/product/components/ProductOptions";
+import ProductQuantity from "@/app/product/components/ProductQuantity";
 import { Button } from "@/components/ui/button";
 import { Form } from "@/components/ui/form";
