+++ app/product/components/ProductDetail.tsx
