+++ app/product/components/ProductNavbar.tsx
