+++ app/product/components/ProductView.tsx
@@ -4,13 +4,13 @@ import { useState } from "react";
 import Image from "next/image";
 import { VisuallyHidden } from "@radix-ui/react-visually-hidden";
 
-import useProductNavigation from "@/app/product/[id]/hooks/useProductNavigation";
+import useProductNavigation from "@/app/product/hooks/useProductNavigation";
 
-import DealTimer from "@/app/(deal)/components/DealTimer";
-import AddToCartForm from "@/app/product/[id]/components/AddToCartForm";
-import ProductDetail from "@/app/product/[id]/components/ProductDetail";
-import ProductNavbar from "@/app/product/[id]/components/ProductNavbar";
-import ProductReview from "@/app/product/[id]/components/ProductReview";
+import DealTimer from "@/app/(main)/deal/components/DealTimer";
+import AddToCartForm from "@/app/product/components/AddToCartForm";
+import ProductDetail from "@/app/product/components/ProductDetail";
+import ProductNavbar from "@/app/product/components/ProductNavbar";
+import ProductReview from "@/app/product/components/ProductReview";
 import { Button } from "@/components/ui/button";
 import { Rating, RatingButton } from "@/components/ui/shadcn-io/rating";
 import {
