+++ app/product/hooks/useProductNavigation.tsx
